requirements.txt
streamlit
pillow
numpy
pydantic
google-generativeai
openai